requirements.txt
# conjunto de paquetes necesario para el tracking
# se deben instalar en la carpeta vendor
# para instalar ejecutar pip install -U -t lib -r requirements.txt

django-filter==0.11
django==1.8.6
djangorestframework==3.3.1
markdown==2.6
sendgrid
simplejson
six
smtpapi
tablib